+++ tests/MessageFactoryTest.php
@@ -5,15 +5,15 @@ declare(strict_types=1);
 namespace Yiisoft\Mailer\Tests;
 
 use InvalidArgumentException;
+use Yiisoft\Mailer\Message;
 use Yiisoft\Mailer\MessageFactory;
-use Yiisoft\Mailer\Tests\TestAsset\DummyMessage;
 
 final class MessageFactoryTest extends TestCase
 {
     public function testCreate(): void
     {
-        $factory = new MessageFactory(DummyMessage::class);
-        $this->assertInstanceOf(DummyMessage::class, $factory->create());
+        $factory = new MessageFactory(Message::class);
+        $this->assertInstanceOf(Message::class, $factory->create());
     }
 
     public function testConstructorThrowExceptionForInvalidMessageClass(): void